Fix stack smash when saving on arm64 (GUID generation)
The DCE UUID code typedef'd unsigned32 as `unsigned long`, which is 8
bytes on 64-bit. uuid_create() casts it over 4-byte buffers (the uuid
time_low field, the 6-byte node id, the 16-byte FGuid), so on arm64 it
wrote past a 6-byte stack array and tripped the stack protector, aborting
every SavePackage/SaveGame. Use `unsigned int` (4 bytes on both ABIs).

+++ app/src/main/cpp/UE1/Source/Core/Src/UnGUID.cpp
@@ -34,7 +34,11 @@
 	Private definitions.
 ----------------------------------------------------------------------------*/
 
-typedef unsigned long   unsigned32;
+// unsigned32 must be exactly 32 bits: this DCE UUID code casts it over 4-byte
+// buffers (uuid_t::time_low, the 6-byte node id, FGuid). `unsigned long` is 8
+// bytes on LP64 (arm64), which overruns those buffers and smashes the stack.
+// `unsigned int` is 4 bytes on both armeabi-v7a and arm64.
+typedef unsigned int    unsigned32;
 typedef unsigned short  unsigned16;
 typedef unsigned char   unsigned8;
 typedef unsigned char   byte;
